refactor: deduplicate auto-discovery and improve repository management
